+++ NEWS.md
@@ -6,6 +6,9 @@
 - ADDED
     - `options_sift()` prints the status of all options when invoked with no arguments. Closes #12.
 
+- FIXED
+    - sift() returns a dataframe of all results, not just the first `n = sift_limit` results.
+
 
 # sift 0.1.0 (2023-03-18)