+++ NEWS.md
@@ -3,6 +3,9 @@
 - ADDED
     - `options_sift()` gets a new option: `sift_peeklength`. This controls the approximate length of the `rand_unique` entries in the data dictionary, i.e. a list of unique values in each column. This "full peek" is used as part of the "haystack" that actually gets searched by `sift()`. It defaults to 3000 characters, but the final length increases when separators are added. Previously, a length limit of only 500 characters was hard-coded in. 3000 characters is about the length of a 1-page Word document at default settings.
 
+- FIXED
+    - `has_class()` can deal with multi-classed variables now.
+
 - CHANGED
     - `some_uniques()` has short-circuit routes for datatypes that don't need the full "random sampling to get a list of its unique values" treatment. So far this is: Factors, Logicals, and Numerics.
     - Changed "peek" separator to vertical bar `|` from comma `,` because some data may use commas within string values.